Fixed Integer Overflow ending in Parse Error
Fixes #18

+++ app/src/main/java/studip_uni_passau/femtopedia/de/unipassaustudip/activities/ScheduleActivity.java
@@ -276,7 +276,7 @@ public class ScheduleActivity extends AppCompatActivity
                             if (Integer.parseInt(time2) == s.getEnd()) {
                                 flag = true;
                                 se.description = s.getContent();
-                                se.color = Integer.parseInt("ff" + s.getColor(), 16);
+                                se.color = (int) Long.parseLong("ff" + s.getColor(), 16);
                             }
                         }
                     }
